create validateBody
