+++ controllers/contacts/contacts.js
@@ -1,13 +1,13 @@
-const Joi = require("joi");
+// const Joi = require("joi");
 
 const contacts = require("../../models/contacts");
 const { HttpError, ctrlWrapper } = require("../../helpers");
 
-const addSchema = Joi.object({
-  name: Joi.string().required(),
-  email: Joi.string().required(),
-  phone: Joi.string().required(),
-});
+// const addSchema = Joi.object({
+//   name: Joi.string().required(),
+//   email: Joi.string().required(),
+//   phone: Joi.string().required(),
+// });
 
 const getAllContacts = async (req, res) => {
   const result = await contacts.listContacts();
@@ -23,10 +23,10 @@ const getById = async (req, res) => {
 };
 
 const addContact = async (req, res) => {
-  const { error } = addSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, "missing required name field");
-  }
+  // const { error } = addSchema.validate(req.body);
+  // if (error) {
+  //   throw HttpError(400, "missing required name field");
+  // }
   const result = await contacts.addContact(req.body);
   res.status(201).json(result);
 };
@@ -41,10 +41,10 @@ const deleteContactById = async (req, res) => {
   });
 };
 const updateContactById = async (req, res) => {
-  const { error } = addSchema.validate(req.body);
-  if (error) {
-    throw HttpError(400, "missing fields");
-  }
+  // const { error } = addSchema.validate(req.body);
+  // if (error) {
+  //   throw HttpError(400, "missing fields");
+  // }
   const { contactId } = req.params;
   const result = await contacts.updateContact(contactId, req.body);
   if (!result) {
