+++ middlewares/index.js
@@ -1,3 +1,5 @@
-// const validateBody = require("./validateBody");
+const validateBody = require("./validateBody");
 
-// module.exports = validateBody;
+module.exports = {
+    validateBody
+};
