+++ middlewares/validateBody.js
@@ -1,14 +1,14 @@
-// const {HttpError} = require("../helpers");
+const {HttpError} = require("../helpers");
 
-// const validateBody = schema => {
-//     const func = (req, res, next)=> {
-//         const { error } = schema.validate(req.body);
-//         if (error) {
-//           next(HttpError(error, error.message));
-//         }
-//         next();
-//       };
-//       return func;
-//     };
+const validateBody = schema => {
+    const func = (req, res, next)=> {
+        const { error } = schema.validate(req.body);
+        if (error) {
+          next(HttpError(400, error.message));
+        }
+        next();
+      };
+      return func;
+    };
     
-//     module.exports = validateBody;
+    module.exports = validateBody;
