+++ routes/api/contacts.js
@@ -2,6 +2,8 @@ const express = require("express");
 
 const ctrl = require("../../controllers/contacts/contacts");
 
+const {validateBody}=require("../../middlewares");
+const schema = require("../../schemas/contacts");
 
 const router = express.Router();
 
@@ -9,10 +11,10 @@ router.get("/", ctrl.getAllContacts);
 
 router.get("/:contactId", ctrl.getById);
 
-router.post("/",  ctrl.addContact);
+router.post("/", validateBody(schema.addSchema), ctrl.addContact);
 
 router.delete("/:contactId", ctrl.deleteContactById);
 
-router.put("/:contactId", ctrl.updateContactById);
+router.put("/:contactId",  validateBody(schema.addSchema), ctrl.updateContactById);
 
 module.exports = router;
